+++ windows/src/common/jwbuiltin.h
@@ -22,14 +22,14 @@
 struct jfile {
 	FILE *f;		/* Regular file, or NULL for built-in */
 	char *p;		/* Built-in file pointer */
-	unsigned int sz;
+	size_t sz;
 	void *dealloc;
 };
 
 typedef struct jfile JFILE;
 
 JFILE *jwfopen(wchar_t *name, wchar_t *mode);
-char *jwfgets(char *buf, int size, JFILE *f);
+char *jwfgets(char *buf, size_t size, JFILE *f);
 int jwfclose(JFILE *f);
 const wchar_t *jwnextbuiltin(const wchar_t *prev, const wchar_t *suffix);
